+++ src/TestTaskStore.ts
@@ -1,6 +1,16 @@
 export class TestTaskStore {
   private static _instance: TestTaskStore;
   private _selectedSchemas: string[] | undefined;
+  private _fileName: string | undefined;
+
+  public get fileName(): string | undefined {
+    return this._fileName;
+  }
+  public set fileName(value: string | undefined) {
+    this._fileName = value;
+  }
+
+
 
   public get selectedSchemas(): string[] | undefined {
     return this._selectedSchemas;