+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+## [0.8.0 - 2021-06-09]
+- Fix wrong error output on package spec when body has errors
+- Execute tests for current package
+
 ## [0.7.0 - 2021-06-09]
 
 - Allow blanklines in SQLPlus